修复缺陷

See merge request nvhl-pcis/nvhl-pcis-ui!777

+++ src/pcis/prodRef/distRef/dist.vue
@@ -167,6 +167,7 @@ const oldPageSchema = ref<any>({});
 const hiddenPage = ref<Array>(['VehicleDist040002']); //初始化需要隐藏的组件
 const addedPlans = ref<string[]>([]);
 const isQuery = ref(false)
+const cProdNos = ['010001','010002','010003','010004','010020'];
 
 watch(
     () => pageresult.list,
@@ -493,7 +494,12 @@ const method = {
           cancelButtonText: "取消",
           type: "warning",
         }
-    ).then(() => {
+    ).then(async () => {
+      // 02大类和01部分产品删除清单时需要先调用保存在执行删除操作，避免清单更新后刷新条款时丢失未保存的条款数据
+      if((route.params.param?.cProdNo.startsWith('02') || cProdNos.includes(route.params.param?.cProdNo)) && !props.compKey?.includes('DeductibleDist')) {
+        const savePlyInfo = await opertaor.getFatherPage().savePlyInfo();
+        if(!savePlyInfo) return;
+      }
       deleteDist(param).then((res: any) => {
         if (res.code === 200) {
           ElMessage.success("删除成功");
@@ -915,7 +921,7 @@ const method = {
         const file = input.files[0];
         const reader = new FileReader();
 
-        reader.onload = (e) => {
+        reader.onload = async (e) => {
           const base64String = e.target?.result as string;
 
           // ✅ 此处赋值有效
@@ -935,6 +941,11 @@ const method = {
             params['cAppNo'] = opertaor.getDataAll().plyBase["Base.cAppNo"]
           }
           
+          // 02大类和01部分产品导入清单时需要先调用保存再执行操作，避免清单更新后刷新条款时丢失未保存的条款数据
+          if((route.params.param?.cProdNo.startsWith('02') || cProdNos.includes(route.params.param?.cProdNo)) && !props.compKey?.includes('DeductibleDist')) {
+            const savePlyInfo = await opertaor.getFatherPage().savePlyInfo();
+            if(!savePlyInfo) return;
+          }
           policyService.importDist(params).then((res) => {
             if (res.code === 200) {
               titleInfo.value = {
@@ -989,7 +1000,7 @@ const method = {
         const file = input.files[0];
         const reader = new FileReader();
 
-        reader.onload = (e) => {
+        reader.onload = async (e) => {
           const base64String = e.target?.result as string;
 
           // ✅ 此处赋值有效
@@ -1007,7 +1018,11 @@ const method = {
           if(route.params.param?.pageName === "priceInquiry") {
             params['cInquiryNo'] = opertaor.getDataAll().plyBase["Base.cInquiryNo"]
           }
-
+          // 02大类和01部分产品导入清单时需要先调用保存再执行操作，避免清单更新后刷新条款时丢失未保存的条款数据
+          if((route.params.param?.cProdNo.startsWith('02') || cProdNos.includes(route.params.param?.cProdNo)) && !props.compKey?.includes('DeductibleDist')) {
+            const savePlyInfo = await opertaor.getFatherPage().savePlyInfo();
+            if(!savePlyInfo) return;
+          }
           policyService.importDistIncrement(params).then((res) => {
             
             if (res.code === 200) {
@@ -1144,6 +1159,11 @@ const method = {
       param['cInquiryNo'] = opertaor.getDataAll().plyBase["Base.cInquiryNo"]
     } else {
       param['cAppNo'] = opertaor.getDataAll().plyBase["Base.cAppNo"]
+    }
+    // 02大类和01部分产品删除清单时需要先调用保存再执行操作，避免清单更新后刷新条款时丢失未保存的条款数据
+    if((route.params.param?.cProdNo.startsWith('02') || cProdNos.includes(route.params.param?.cProdNo)) && !props.compKey?.includes('DeductibleDist')) {
+      const savePlyInfo = await opertaor.getFatherPage().savePlyInfo();
+      if(!savePlyInfo) return;
     }
     // 增加是否删除全部判断，选是删除全部，选否删除选中项
     if (pageresult.total > 10 && selectedRows.value.length === 10) {
@@ -1315,16 +1335,12 @@ const getSummary = async () => {
 async function refreshCvrg() {
   // 如果是免赔信息则不刷新保障信息
   if(props.compKey?.includes('DeductibleDist')) return;
-  const cProdNos = ['010001','010002','010003','010004','010020'];
   if(route.params.param?.cProdNo.startsWith('02') || cProdNos.includes(route.params.param?.cProdNo)) {
     const cvrgRef = opertaor.getTableRefs()['cvrg'];
-    const savePlyInfo = await opertaor.getFatherPage().savePlyInfo();
-    if(savePlyInfo) {
-      try {
-        cvrgRef?.getAddrSeqOptions();
-        cvrgRef?.refushCvrgInfo();
-      } catch (ignore) {
-      }
+    try {
+      cvrgRef?.getAddrSeqOptions();
+      cvrgRef?.refushCvrgInfo();
+    } catch (ignore) {
     }
   }
 }
